+++ src/TCP/TCP.cs
@@ -1,7 +1,0 @@
-class TCP
-{
-    public static void Print()
-    {
-        Console.WriteLine("Hello World!");
-    }
-}