+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## 9.2.2025
 - Structured the directory, created makefile
+- Added some base class skeletons. Added regular expressions for the TCP messages according to the assignment grammar.